requirements.txt
anyio==4.0.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
inflection==0.5.1
packaging==23.2
PyJWT==1.7.1
pyTelegramBotAPI==4.14.0
python-decouple==3.8
python-telegram-bot==20.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
